Discovery server now communicates with the other services
Added dependencies in pom for services, as well as eureka server for discovery server and eureka client for the services. Updated todo.md to remember not to hardcode port adresses. The services can now run in and have connection to the eureka server. The services can also have multiple instances.